+++ packages/types/src/types/registry.ts
@@ -3,6 +3,7 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 import { ChainProperties } from '../interfaces/rpc';
+import { OverrideModuleType } from '../known/types';
 
 import BN from 'bn.js';
 
@@ -81,10 +82,31 @@ export interface RegistryMetadata {
   asLatest: RegistryMetadataLatest;
 }
 
+export interface RegisteredTypes {
+  /**
+   * @description Additional types used by runtime modules. This is necessary if the runtime modules
+   * uses types not available in the base Substrate runtime.
+   */
+  types?: RegistryTypes;
+  /**
+   * @description Alias an types, as received via the metadata, to a JS-specific type to avoid conflicts. For instance, you can rename the `Proposal` in the `treasury` module to `TreasuryProposal` as to not have conflicts with the one for democracy.
+   */
+  typesAlias?: Record<string, OverrideModuleType>;
+  /**
+   * @description Additional types that are injected based on the chain we are connecting to. There are keyed by the chain, i.e. `{ 'Kusama CC1': { ... } }`
+   */
+  typesChain?: Record<string, RegistryTypes>;
+  /**
+   * @description Additional types that are injected based on the type of node we are connecting to, as set via specName in the runtime version. There are keyed by the node, i.e. `{ 'edgeware': { ... } }`
+   */
+  typesSpec?: Record<string, RegistryTypes>;
+}
+
 export interface Registry {
   readonly chainDecimals: number;
   readonly chainSS58: number | undefined;
   readonly chainToken: string;
+  readonly knownTypes: RegisteredTypes;
 
   findMetaCall (callIndex: Uint8Array): CallFunction;
   findMetaError (errorIndex: Uint8Array): any;
@@ -99,6 +121,7 @@ export interface Registry {
   getDefinition (name: string): string | undefined;
   getOrThrow <T extends Codec = Codec> (name: string, msg?: string): Constructor<T>;
   getOrUnknown <T extends Codec = Codec> (name: string): Constructor<T>;
+  setKnownTypes (types: RegisteredTypes): void;
   getSignedExtensionExtra (): Record<string, keyof InterfaceTypes>;
   getSignedExtensionTypes (): Record<string, keyof InterfaceTypes>;
   hasClass (name: string): boolean;